requirements.txt
absl-py==1.4.0
anyio==3.7.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.4
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.97.0
filelock==3.12.2
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h11==0.14.0
h5py==3.8.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
jax==0.4.12
Jinja2==3.1.2
keras==2.12.0
libclang==16.0.0
lit==16.0.5.post0
Markdown==3.4.3
MarkupSafe==2.1.3
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.9.1
packaging==23.1
Pillow==9.5.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-addons==0.20.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
torch==2.0.1
torchvision==0.15.2
triton==2.0.0
typeguard==2.13.3
typing_extensions==4.6.3
ujson==5.8.0
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.14.1
zipp==3.15.0
